+++ tests/compilation_failures/core/extend_flattened_variable.rs
@@ -3,6 +3,6 @@ use preinterpret::*;
 fn main() {
     preinterpret! {
         [!set! #variable = 1]
-        [!extend! #..variable += 2]
+        [!set! #..variable += 2]
     }
 }