+++ tests/compilation_failures/core/extend_non_existing_variable.rs
@@ -2,6 +2,6 @@ use preinterpret::*;
 
 fn main() {
     preinterpret! {
-        [!extend! #variable += 2]
+        [!set! #variable += 2]
     }
 }