+++ tests/compilation_failures/core/extend_variable_and_then_extend_it_again.rs
@@ -3,6 +3,6 @@ use preinterpret::*;
 fn main() {
     preinterpret! {
         [!set! #variable = Hello]
-        [!extend! #variable += World [!extend! #variable += !]]
+        [!set! #variable += World [!set! #variable += !]]
     }
 }