+++ tests/compilation_failures/core/extend_variable_and_then_read_it.rs
@@ -3,6 +3,6 @@ use preinterpret::*;
 fn main() {
     preinterpret! {
         [!set! #variable = Hello]
-        [!extend! #variable += World #variable]
+        [!set! #variable += World #variable]
     }
 }