+++ tests/compilation_failures/core/extend_variable_and_then_set_it.rs
@@ -3,6 +3,6 @@ use preinterpret::*;
 fn main() {
     preinterpret! {
         [!set! #variable = Hello]
-        [!extend! #variable += World [!set! #variable = Hello2]]
+        [!set! #variable += World [!set! #variable = Hello2]]
     }
 }